Rename cumd to cpumond
