+++ Makefile.am
@@ -6,7 +6,7 @@
 SUBDIRS  = lvm
 SUBDIRS += $(MAYBE_part)
 SUBDIRS += vhd
-SUBDIRS += cumd
+SUBDIRS += cpumond
 SUBDIRS += thin
 SUBDIRS += control
 SUBDIRS += drivers
